feat(#561): Loading screen: splash component exists but never integrated into extension loading
